feat: display auth errors
