+++ public/locales/en/translation.json
@@ -14,5 +14,13 @@
   "Valid email required.": "Valid email required.",
   "Valid password required.": "Valid password required.",
   "Min length of password is 8 symbols": "Min length of password is 8 symbols",
-  "Password should contain minimum 8 symbols, at least one letter, one digit, one special character": "Password should contain minimum 8 symbols, at least one letter, one digit, one special character($%_+!-)"
+  "Password should contain minimum 8 symbols, at least one letter, one digit, one special character": "Password should contain minimum 8 symbols, at least one letter, one digit, one special character($%_+!-)",
+  "email-already-exists": "Email already exists",
+  "internal-error": "Auth server internal error",
+  "invalid-email": "Invalid email",
+  "invalid-password": "Invalid password",
+  "wrong-password": "Wrong password",
+  "user-not-found": "User not found",
+  "unknown error": "Unknown error",
+  "too-many-requests": "Limit of requests exceeded"
 }
