+++ src/index.css
@@ -74,3 +74,8 @@ button:focus-visible {
     background-color: #f9f9f9;
   }
 }
+.center {
+  position: absolute;
+  top: 40vh;
+  left: 46vw;
+}
